+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Reader**) Open the first page when going to the previous chapter
 - (**Download**) Group downloads by source
 - (**Download**) Add series bulk actions
+- (**Download**) Make downloads cancelable outside the download queue page
 
 ### Fixed